\name{arm-package}
\alias{arm-package}
\alias{arm}

\docType{package}

\title{Data Analysis Using Regression and Multilevel/Hierarchical Models}

\description{R functions for processing lm, glm, mer and polr outputs.}

\details{\tabular{ll}{Package:\tab arm\cr
Version:\tab 1.0-26\cr
Date:\tab 2007/08/13\cr
Namespace:\tab arm\cr
Depends:\tab methods, R (>= 2.4.0), MASS, Matrix (>= 0.9975-1),\cr
        \tab lme4 (>= 0.9975-1), R2WinBUGS\cr
Suggest:\tab car, foreign, nnet\cr
License:\tab GPL version 2 or newer\cr}
}

\author{\tabular{l}{Andrew Gelman <gelman@stat.columbia.edu>,\cr
        Yu-Sung Su <ys463@columbia.edu>,\cr
        Jennifer Hill <jh1030@columbia.edu>,\cr
        Maria Grazia Pittau <grazia@stat.columbia.edu>,\cr
        Jouni Kerman <jouni@kerman.com> and\cr
        Tian Zheng <tzheng@stat.columbia.edu>\cr}
}

\references{Gelman, Andrew and Jennifer Hill. (2007). 
            \emph{Data Analysis Using Regression and Multilevel/Hierarchical Models}. 
            NY: Cambridge University Press.\url{http://www.stat.columbia.edu/~gelman/arm/}.}

\note{Maintainer: Yu-Sung Su <ys463@columbia.edu>}

\keyword{package}
