% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/contrasts.bayes.R
\name{contrasts.bayes}
\alias{contr.bayes.ordered}
\alias{contr.bayes.unordered}
\alias{contrasts.bayes}
\title{Contrast Matrices}
\usage{
contr.bayes.ordered(n, scores = 1:n, contrasts = TRUE)

contr.bayes.unordered(n, base = 1, contrasts = TRUE)
}
\arguments{
\item{n}{a vector of levels for a factor, or the number of levels.}

\item{scores}{the set of values over which orthogonal polynomials
are to be computed.}

\item{contrasts}{a logical indicating whether contrasts should be computed.}

\item{base}{an integer specifying which group is considered the
baseline group. Ignored if \code{contrasts} is \code{FALSE}.}
}
\description{
Return a matrix of contrasts used in \code{\link{bayesglm}}.
}
\details{
These functions are adapted from \code{contr.treatment} and
\code{contr.poly} in \code{\link{stats}} package.  The purpose for
these functions are to keep the baseline levels of categorical
variables and thus to suit the use of \code{\link{bayesglm}}.

\code{contr.bayes.unordered} is equivalent to
\code{contr.treatment} whereas \code{contr.bayes.ordered} is
equivalent to \code{contr.poly}.
}
\examples{
cat.var <- rep(1:3, 5)
dim(contr.bayes.unordered(cat.var))
# 15*15 baseline level kept!
dim(contr.treatment(cat.var))
# 15*14
}
\author{
Yu-Sung Su \email{suyusung@tsinghua.edu.cn}
}
\seealso{
\code{\link{C}}, \code{\link{contr.helmert}},
\code{\link{contr.poly}}, \code{\link{contr.sum}},
\code{\link{contr.treatment}}; \code{\link{glm}},
\code{\link{aov}}, \code{\link{lm}}, \code{\link{bayesglm}}.
}
\keyword{array}
\keyword{design}
\keyword{manip}
\keyword{regression}

