% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2cov.R
\name{cor2cov}
\alias{cor2cov}
\title{Compute covariance matrix}
\usage{
cor2cov(cor, sd)
}
\arguments{
\item{cor}{a correlation matrix}

\item{sd}{a vector of standard deviations}
}
\value{
a variance covariance matrix
}
\description{
This function computes a covairance matrix based on a correlation matrix and
a vector of standard deviations
}
\examples{
cor_mat <- rbind(c( 1,   0,  .3),
                 c( 0,   1, -.2),
                 c(.3, -.2,   1))

SDs <- c(5, 10, 3)
cor2cov(cor_mat, sd = SDs)

}
