% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_results}
\alias{extract_results}
\title{Extract results}
\usage{
extract_results(x)
}
\arguments{
\item{x}{object of class `linearHypothesis.mlm`}
}
\value{
An object of class `c("anova", "data.frame")` with the test information.
}
\description{
This function extracts the test results from linear hypothesis tests
on multivariate linear models as returned by the `car` package
}
\details{
This function is a copy of the S3 `print()` method for objects of the
`linearHypothesis.mlm`-class. However, rather than invisibly returning
the object, the printed test results are returned as a data.frame
}
