% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extractDIC.R
\name{extractDIC}
\alias{extractDIC}
\alias{extractDIC.merMod}
\title{Extract AIC and DIC from a mer model}
\usage{
extractDIC(fit, ...)

\method{extractDIC}{merMod}(fit, ...)
}
\arguments{
\item{fit}{fitted \code{merMod} mode, usually the result of a
fitter like \code{merMod}.}

\item{...}{further arguments (currently unused).}
}
\description{
Computes the (generalized) Akaike *A*n *I*nformation *C*riterion
and *D*eviance *I*nformation *C*riterion for a mer model.
}
\examples{
fm1 <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy)
extractAIC(fm1)
extractDIC(fm1)
}
\author{
Andrew Gelman \email{gelman@stat.columbia.edu};
  Yu-Sung Su \email{suyusung@tsinghua.edu.cn}
}
\keyword{manip}
\keyword{methods}

