% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fround.R
\name{fround}
\alias{fround}
\alias{pfround}
\title{Formating the Rounding of Numbers}
\usage{
fround(x, digits)

pfround(x, digits)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{integer indicating the precision to be used.}
}
\description{
\code{fround} rounds the values in its first argument to the
  specified number of decimal places with surrounding quotes.

  \code{pfround} rounds the values in its first argument to the
  specified number of decimal places without surrounding quotes.
}
\examples{
x <- rnorm(1)
fround(x, digits=2)
pfround(x, digits=2)
}
\author{
Andrew Gelman \email{gelman@stat.columbia.edu};
  Yu-Sung Su \email{suyusung@tsinghua.edu.cn}
}
\seealso{
\code{\link{round}}
}
\keyword{manip}
\keyword{print}

