% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/go.R
\docType{data}
\name{go}
\alias{G}
\alias{go}
\alias{print.GO}
\title{Function to Recall Last Source File}
\format{\preformatted{Class 'GO'  logi NA
}}
\usage{
go(..., add = FALSE, timer = FALSE)

print.GO(x, ...)

G
}
\arguments{
\item{add}{add these names to the current list; if replace, then
\code{FALSE}.}

\item{timer}{time the execution time of \code{go()}.}

\item{\dots}{list of filenames as character strings.}
}
\description{
A function that like \code{source()} but recalls the last source
file names by default.
}
\examples{
go('myprog')            # will run source('myprog.r')
go()                    # will run source('myprog.r') again
go('somelib',add=TRUE)  # will run source('myprog.r') and source('somelib.r')
go('myprog','somelib')  # same as above
go('mytest')            # will run source('mytest') only
go()                    # runs source('mytest') again
G                       # short cut to call go()
}
\author{
Jouni Kerman \email{jouni@kerman.com}
\email{kerman@stat.columbia.edu}
}
\keyword{datasets}
\keyword{manip}
\keyword{methods}

