% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/invlogit.R
\name{invlogit}
\alias{invlogit}
\alias{logit}
\title{Logistic and Inverse logistic functions}
\usage{
invlogit(x)

logit(x)
}
\arguments{
\item{x}{A vector of continuous values.}
}
\value{
A vector of estimated probabilities.
}
\description{
Inverse-logit function, transforms continuous values to the range
(0, 1)
}
\details{
The Inverse-logit function defined as: \eqn{logit^-1(x) =
e^x/(1+e^x)} transforms continuous values to the range (0, 1),
which is necessary, since probabilities must be between 0 and 1 and
maps from the linear predictor to the probabilities
}
\examples{
data(frisk)
n <- 100
x1 <- rnorm (n)
x2 <- rbinom (n, 1, .5)
b0 <- 1
b1 <- 1.5
b2 <- 2
Inv.logit <- invlogit(b0+b1*x1+b2*x2)
plot(b0+b1*x1+b2*x2, Inv.logit)
}
\author{
Andrew Gelman \email{gelman@stat.columbia.edu},
  M.Grazia Pittau \email{grazia@stat.columbia.edu}
}
\references{
Andrew Gelman and Jennifer Hill. (2006). \emph{Data
  Analysis Using Regression and Multilevel/Hierarchical
  Models}. Cambridge University Press.
}
\keyword{models}

