\name{readColumns}
% functions
\alias{read.columns}

\title{Function to read data by columns}

\description{
    A function read data by columns
}
\usage{
read.columns(filename, columns)
}
\arguments{
  \item{filename}{user specified file name including path of the file}
  \item{columns}{which columns of the data to be read}
}



\author{Andrew Gelman \email{gelman@stat.columbia.edu}
}



\keyword{methods}
