% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model.matrix.bayes.R
\name{model.matrixBayes}
\alias{model.matrixBayes}
\title{Construct Design Matrices}
\usage{
model.matrixBayes(object, data = environment(object), contrasts.arg = NULL,
  xlev = NULL, keep.order = FALSE, drop.baseline = FALSE, ...)
}
\description{
\code{model.matrixBayes} creates a design matrix.
}
\details{
\code{model.matrixBayes} is adapted from \code{model.matrix} in the
\code{stats} pacakge and is designed for the use of
\code{\link{bayesglm}}.% and \code{bayesglm.hierachical} (not yet
implemented!). It is designed to keep baseline levels of all
categorical varaibles and keep the variable names unodered in the
output.  The design matrices created by \code{model.matrixBayes}
are unidentifiable using classical regression methods, though; they
can be identified using \code{\link{bayesglm}}.% and
%\code{bayesglm.hierachical}.
}
\examples{
ff <- log(Volume) ~ log(Height) + log(Girth)
str(m <- model.frame(ff, trees))
(model.matrix(ff, m))
class(ff) <- c("bayesglm", "terms", "formula")
(model.matrixBayes(ff, m))
\%class(ff) <- c("bayesglm.h", "terms", "formula")
\%(model.matrixBayes(ff, m))
}
\author{
Yu-Sung Su \email{suyusung@tsinghua.edu.cn}
}
\references{
Andrew Gelman, Aleks Jakulin, Maria Grazia Pittau and
Yu-Sung Su. (2009). \dQuote{A Weakly Informative Default Prior
Distribution For Logistic And Other Regression Models.} \emph{The
Annals of Applied Statistics} 2 (4):
1360--1383. \url{http://www.stat.columbia.edu/~gelman/research/published/priors11.pdf}
}
\seealso{
\code{\link[stats]{model.frame}},
\code{\link[stats]{model.extract}}, \code{\link[stats]{terms}},
\code{\link[stats]{terms.formula}}, \code{\link{bayesglm}}.
}
\keyword{manip}
\keyword{models}

